Implement repair
Now that we have a notion of a purged block, we can "repair" a tree by
just purging everything that can not be resolved or decoded. Of course we
lose data, but it beats losing the entire tree.

Purging non-sealed data is a bit hairy though, and repairing can obviously
leave you with gaps instead of just a bit missing in front.